+++ src/includes/classes/App.php
@@ -183,7 +183,7 @@ class App extends Classes\Core\Base\Core
      *
      * @var string Version.
      */
-    const VERSION = '161013.18318'; //v//
+    const VERSION = '161204.42372'; //v//
 
     /**
      * Constructor.